0x02-functions, nested loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 /**
  * Main - prints the first 50 Fibinacci numbers
- * Return: 0 (Sucess)
+ * Return: 0 (Success)
  */
 int main(void)
 {
 	int c;
-	int f1 = 0, f2 = 0;
+	unsigned long f1 = 0, f2 = 0;
 	int sum;
 
 	for (c = 0; c < 50; c++)
 	{
 		sum = f1 + f2;
-		printf("%d", sum);
+		printf("%lu", sum);
 		f1 = f2;
 		f2 = sum;
